Move index logic to Loader component
Also add back some commented tests in other places
