+++ packages/jbrowse-desktop/src/Loader.tsx
@@ -1,0 +1,99 @@
+import PluginManager from '@gmod/jbrowse-core/PluginManager'
+import CircularProgress from '@material-ui/core/CircularProgress'
+import { makeStyles } from '@material-ui/core/styles'
+import React, { useEffect, useState } from 'react'
+import corePlugins from './corePlugins'
+import JBrowse from './JBrowse'
+import JBrowseRootModelFactory from './rootModel'
+
+const { electron } = window
+
+const useStyles = makeStyles({
+  loadingIndicator: {
+    position: 'fixed',
+    top: '50%',
+    left: '50%',
+    marginTop: -25,
+    marginLeft: -25,
+  },
+})
+
+export default function Loader() {
+  const [configSnapshot, setConfigSnapshot] = useState()
+  const [plugins, setPlugins] = useState()
+
+  const classes = useStyles()
+
+  const ipcRenderer = (electron && electron.ipcRenderer) || {
+    invoke: () => {},
+  }
+
+  useEffect(() => {
+    async function fetchConfig() {
+      try {
+        setConfigSnapshot(
+          Object.assign(await ipcRenderer.invoke('loadConfig'), {
+            configuration: { rpc: { defaultDriver: 'ElectronRpcDriver' } },
+          }),
+        )
+      } catch (error) {
+        setConfigSnapshot(() => {
+          throw error
+        })
+      }
+    }
+
+    fetchConfig()
+  }, [ipcRenderer])
+
+  useEffect(() => {
+    async function fetchPlugins() {
+      // TODO: Runtime plugins
+      // Loading runtime plugins will look something like this
+      // const pluginLoader = new PluginLoader(config.plugins)
+      // const runtimePlugins = await pluginLoader.load()
+      // setPlugins([...corePlugins, ...runtimePlugins])
+      setPlugins(corePlugins)
+    }
+    fetchPlugins()
+  }, [])
+
+  if (!(configSnapshot && plugins)) {
+    return (
+      <CircularProgress
+        disableShrink
+        className={classes.loadingIndicator}
+        size={50}
+      />
+    )
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const pluginManager = new PluginManager(plugins.map((P: any) => new P()))
+  pluginManager.createPluggableElements()
+
+  const JBrowseRootModel = JBrowseRootModelFactory(pluginManager)
+  let rootModel
+  try {
+    if (configSnapshot) {
+      rootModel = JBrowseRootModel.create({ jbrowse: configSnapshot })
+    }
+  } catch (e) {
+    console.error(e)
+    const additionalMsg =
+      e.message.length > 10000 ? '... see console for more' : ''
+    throw new Error(e.message.slice(0, 10000) + additionalMsg)
+  }
+
+  // make some things available globally for testing
+  // e.g. window.MODEL.views[0] in devtools
+  // @ts-ignore
+  window.MODEL = rootModel.session
+  // @ts-ignore
+  window.ROOTMODEL = rootModel
+  pluginManager.setRootModel(rootModel)
+
+  pluginManager.configure()
+
+  return <JBrowse pluginManager={pluginManager} />
+}
