+++ packages/jbrowse-web/src/declare.d.ts
@@ -2,3 +2,4 @@ declare module '@gmod/jbrowse-core/configuration'
 declare module '@gmod/jbrowse-core/configuration/configurationSchema'
 declare module '@gmod/jbrowse-core/util/tracks'
 declare module '@gmod/jbrowse-core/PluginManager'
+declare module 'fastestsmallesttextencoderdecoder'
